requirements.txt
flask==3.0.0
flask-cors==4.0.0
selenium==4.15.2
webdriver-manager==4.0.1
gunicorn==21.2.0
python-dotenv==1.0.0

